refactor : 이메일인증 방식 변경
