+++ src/main/java/com/example/be12th/global/config/SecurityConfig.java
@@ -29,7 +29,7 @@ public class SecurityConfig {
 
                         //user
                         .requestMatchers(HttpMethod.POST,"/user/email").permitAll()
-                        .requestMatchers(HttpMethod.POST,"/user/email/verify").permitAll()
+                        .requestMatchers(HttpMethod.POST,"/user/verify/signup").permitAll()
                         .requestMatchers(HttpMethod.POST,"/user/signup").permitAll()
                         .requestMatchers(HttpMethod.POST,"/user/login").permitAll()
                         .anyRequest().authenticated()
